Add Home component setup
