+++ src/components/Home/Home.js
@@ -1,0 +1,9 @@
+import './Home.css'
+
+export default function Home() {
+    return (
+        <div className="container">
+
+        </div>
+    )
+}
