+++ src/components/Liquidity/CakeLP-OWN/CakeLP_OWN.css
@@ -1,0 +1,68 @@
+#app-staking {
+  margin-top: 80px;
+  padding: 5rem 0;
+}
+
+.app-card {
+  background-color: white;
+  border-radius: 25px;
+  box-shadow: 0px 0px 25px -10px rgba(0, 0, 0, 0.25);
+  padding: 1.5em;
+}
+
+.total-dep {
+  padding: 0.75em;
+  width: 70%;
+  margin: 0 auto;
+}
+
+.stake-form {
+  position: relative;
+  margin-bottom: 0.75rem;
+}
+
+.stake-input {
+  font-size: 0.9rem;
+  padding: 0.85rem 1rem;
+  margin-bottom: 0.25rem;
+}
+
+.stake-btn {
+  background-color: rgba(22, 185, 154, 0.1);
+  color: rgb(22, 185, 154);
+  padding: 6px 14px;
+  border-radius: 6px;
+  position: absolute;
+  right: 7px;
+  top: 6px;
+}
+
+.stake-btn:hover,
+.stake-btn:focus {
+  background-color: rgba(22, 185, 154, 0.15);
+  color: rgb(22, 185, 154);
+}
+
+.stake-btn-func {
+  width: 49%;
+}
+
+.stake-link {
+  color: #000000;
+}
+
+.stake-link:hover,
+.stake-link:focus {
+  color: #16b99a;
+}
+
+/* Remove arrows in number input */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+
+input[type="number"] {
+  -moz-appearance: textfield;
+}
